Code optimization and cleanup
