+++ src/components/IntroSection/style.css
@@ -42,7 +42,6 @@
 .intro-description {
 	margin: 5px;
 	font-size: 2.75rem;
-	/* font-size: 5rem; */
 	padding: 0;
 	color: var(--primary-theme-color);
 }
