+++ src/components/Testimonials/style.css
@@ -76,7 +76,7 @@
 
 q.highlight::before {
 	content: open-quote;
-	font-size: 6rem;
+	font-size: 4rem;
 	font-family: serif;
 	color: var(--about-section-testimonial-quote-color);
 }
@@ -130,5 +130,9 @@ q.highlight::after,
 	.highlight {
 		font-size: 1.5rem;
 		line-height: 32px;
+	}
+
+	q.highlight::before {
+		font-size: 6rem;
 	}
 }
