+++ src/components/common/LoadingSpinner/LoadingSpinner.module.css
@@ -1,0 +1,4 @@
+.root > div {
+	margin: 5rem auto;
+	text-align: center;
+}
